+++ tests/Unit/FeaturePolicyTest.php
@@ -14,7 +14,7 @@ final class FeaturePolicyTest extends TestCase {
 		self::assertSame('FALSE', $config['IM']['im_status']);
 		self::assertSame('FALSE', $config['Sms']['sms_status']);
 		self::assertSame('FALSE', $config['Video']['video_enableVideo']);
-		self::assertSame('TRUE', $config['cloudcontacts']['cloudcontacts_status']);
+		self::assertSame('FALSE', $config['cloudcontacts']['cloudcontacts_status']);
 		self::assertSame('FALSE', $config['configurations']['configurations_dndVisibility']);
 		self::assertSame('FALSE', $config['blf']['autoenable_blf']);
 		self::assertSame('FALSE', $config['zoom']['zoom_status']);